Bootstrap Api Setup

+++ app/src/main/java/com/nyotek/dot/admin/models/responses/NSUserResponse.kt
@@ -15,7 +15,7 @@ data class NSUserResponse(
 data class NSDataUser(
     @SerializedName("expiresIn")
     @Expose
-    var expiresIn: Double? = null,
+    var expiresIn: Long? = null,
     @SerializedName("accessToken")
     @Expose
     var accessToken: String? = null,
